+++ tools/flang2/flang2exe/llutil.cpp
@@ -2062,7 +2062,7 @@ write_constant_value(int sptr, LL_Type *type, INT conval0, INT conval1,
     if ((num[0] & 0x7ff00000) == 0x7ff00000) /* exponent == 2047 */
       sprintf(d, "0x%08x%08x", num[0], num[1]);
     /* also check for -0 */
-    else if (num[0] == 0x80000000 && num[1] == 0x00000000)
+    else if (num[0] == (INT)0x80000000 && num[1] == (INT)0x00000000)
       sprintf(d, "-0.00000000e+00");
     /* remember to make room for /0 */
     fprintf(LLVMFIL, "%s", d);
@@ -2093,7 +2093,7 @@ write_constant_value(int sptr, LL_Type *type, INT conval0, INT conval1,
       sprintf(constant2, "%X", dtmp.tmp[0]);
 
     /* check for negative zero */
-    if (dtmp.tmp[1] == 0x80000000 && !dtmp.tmp[0])
+    if (dtmp.tmp[1] == (INT)0x80000000 && !dtmp.tmp[0])
       fprintf(LLVMFIL, "-0.000000e+00");
     else
       fprintf(LLVMFIL, "0x%s%s", constant1, constant2);
